requirements.txt
streamlit
audio_recorder_streamlit
streamlit_float
python-dotenv
groq
boto3
